#166 - Fixed bug with GetTeam() not casting properly - Added simple regex for catching bad markup in chat - Refactoring InRoomChat - Refactoring ChatCommandHandler
