+++ Assets/Scripts/Room/Chat/ChatCommandLineHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5dd4dc0bb6cb436499888dd6aee076f6
+guid: 76e808a9a4d41854f8cc63dc73385447
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
